requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
aiofiles==23.1.0
pydantic==1.10.8
